Use etree to interact with manifest files
libcombine was causing problems with the build (and had some other
quirks while developing... decided it may not be worth using it for
this simple functionality.

Python xml libs are not exactly great (default namespace handling
is awkward, and no pretty printing unless we want to reparse and format
it through minidom) but should be more reliable at least.