Don't pretend we can delete ephemeral files
They should be excluded from the list of existing files shown in the
new version view, since it's not possible actually to delete them:
they won't be retained in the new version anyway. You get a git error
if you try to delete them, since they don't really exist as part of
the commit!