Better approach to selecting visible versions
This is still a WIP, but now works and should be much more efficient
in terms of DB access - I think it should be down to 2 queries,
although I haven't yet checked this live, only running tests.

@helenst, it might be worth a look at this approach in the context
of the extra experiment matrix views (#64) and later for #73? We're
going to need to extend the visibility checks anyway for the former,
and probably lots of places will need to use either my `~Q()` or
`.exclude()` to select visibility != private.