Remove old generic caches
Also update all code using these to the new type-specific caches.

This required a slight hack of adding a `cachedentity` property to
`Entity` that points at the appropriate cache for the `entity_type`;
once we make `Entity` an abstract model itself then this shouldn't
be needed any more, as there will be separate `ModelEntity` and
`ProtocolEntity` tables and so each can have its own `cachedentity` link.